requirements.txt
absl-py==1.4.0
aioboto3==10.4.0
aiobotocore==2.4.2
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
albumentations==1.3.0
arrow==1.2.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
binaryornot==0.4.4
boto3==1.24.59
botocore==1.27.59
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
deeplake==3.2.22
dill==0.3.6
entrypoints==0.4
filelock==3.9.0
flatbuffers==23.3.3
frozenlist==1.3.3
gast==0.4.0
google-auth==2.17.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h5py==3.8.0
humbug==0.3.1
idna==3.4
imageio==2.27.0
importlib-metadata==6.3.0
jax==0.4.8
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
joblib==1.2.0
keras==2.12.0
lazy_loader==0.2
libclang==16.0.0
lit==15.0.7
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
nest-asyncio==1.5.6
networkx==3.0
numcodecs==0.11.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
pathos==0.3.0
Pillow==9.3.0
pox==0.3.2
ppft==1.7.6.6
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.6.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.16.0
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.3.21
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
zipp==3.15.0
